[IC-105] chore: scanSave와 scanHistory 패키지 분리 (#84)
