+++ scanhistory/src/main/java/com/initcloud/rocket23/checklist/entity/scanHistory/ProjectVulnDetail.java
@@ -1,4 +1,4 @@
-package com.initcloud.rocket23.checklist.entity;
+package com.initcloud.rocket23.checklist.entity.scanHistory;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
