+++ scanhistory/src/main/java/com/initcloud/rocket23/infra/repository/CheckovScanRepository.java
@@ -1,6 +1,6 @@
 package com.initcloud.rocket23.infra.repository;
 
-import com.initcloud.rocket23.checklist.entity.CheckovScan;
+import com.initcloud.rocket23.checklist.entity.scanResult.CheckovScan;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
