+++ scanhistory/src/main/java/com/initcloud/rocket23/infra/repository/ProjectVulnDetailRepository.java
@@ -1,6 +1,6 @@
 package com.initcloud.rocket23.infra.repository;
 
-import com.initcloud.rocket23.checklist.entity.ProjectVulnDetail;
+import com.initcloud.rocket23.checklist.entity.scanHistory.ProjectVulnDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
