+++ scanhistory/src/main/java/com/initcloud/rocket23/infra/repository/ScanHistoryDetailRepository.java
@@ -1,6 +1,6 @@
 package com.initcloud.rocket23.infra.repository;
 
-import com.initcloud.rocket23.checklist.entity.ScanHistoryDetail;
+import com.initcloud.rocket23.checklist.entity.scanHistory.ScanHistoryDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
