+++ scanhistory/src/main/java/com/initcloud/rocket23/infra/repository/ScanHistoryRepository.java
@@ -2,7 +2,7 @@ package com.initcloud.rocket23.infra.repository;
 
 import java.util.Optional;
 
-import com.initcloud.rocket23.checklist.entity.ScanHistory;
+import com.initcloud.rocket23.checklist.entity.scanHistory.ScanHistory;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -11,9 +11,10 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface ScanHistoryRepository extends JpaRepository<ScanHistory, Long> {
-	Optional<ScanHistory> findTopByTeam_TeamCodeAndProject_ProjectCodeAndFileHashOrderById(String teamCode,
-		String projectCode, String hashCode);
+    Optional<ScanHistory> findTopByTeam_TeamCodeAndProject_ProjectCodeAndFileHashOrderById(String teamCode,
+                                                                                           String projectCode,
+                                                                                           String hashCode);
 
-	Page<ScanHistory> findAllByTeam_TeamCodeAndProject_ProjectCode(Pageable pageable, String teamCode,
-		String projectCode);
+    Page<ScanHistory> findAllByTeam_TeamCodeAndProject_ProjectCode(Pageable pageable, String teamCode,
+                                                                   String projectCode);
 }
