+++ scanhistory/src/main/java/com/initcloud/rocket23/team/entity/Team.java
@@ -1,6 +1,6 @@
 package com.initcloud.rocket23.team.entity;
 
-import com.initcloud.rocket23.checklist.entity.ScanHistory;
+import com.initcloud.rocket23.checklist.entity.scanHistory.ScanHistory;
 import com.initcloud.rocket23.common.entity.BaseEntity;
 import com.initcloud.rocket23.common.utils.UniqueUtils;
 import com.initcloud.rocket23.team.dto.TeamDto;
@@ -16,33 +16,33 @@ import java.util.List;
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Team extends BaseEntity {
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "TEAM_ID")
-	private Long id;
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "TEAM_ID")
+    private Long id;
 
-	@Column
-	private String teamCode;
+    @Column
+    private String teamCode;
 
-	@Column
-	private String name;
+    @Column
+    private String name;
 
-	@Column
-	private String logoUri;
+    @Column
+    private String logoUri;
 
-	@OneToMany(mappedBy = "team")
-	private List<ScanHistory> scanHistories = new ArrayList<>();
+    @OneToMany(mappedBy = "team")
+    private List<ScanHistory> scanHistories = new ArrayList<>();
 
-	public Team(Long id, String teamCode, String name, String logoUri) {
-		this.id = id;
-		this.teamCode = teamCode;
-		this.name = name;
-		this.logoUri = logoUri;
-	}
+    public Team(Long id, String teamCode, String name, String logoUri) {
+        this.id = id;
+        this.teamCode = teamCode;
+        this.name = name;
+        this.logoUri = logoUri;
+    }
 
-	public Team(final TeamDto team) {
-		this.teamCode = UniqueUtils.getUUID();
-		this.name = team.getTeamName();
-		this.logoUri = team.getLogoUri();
-	}
+    public Team(final TeamDto team) {
+        this.teamCode = UniqueUtils.getUUID();
+        this.name = team.getTeamName();
+        this.logoUri = team.getLogoUri();
+    }
 }
